Refactor: [WIP] search 컴포넌트 정리
