--- components/search/SearchOptions.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import SearchOptionsComponent from '@/components/search/SearchOptionsComponent';
-
-export default function SearchOptions() {
-  // const { tab, setTab } = useTabBar();
-
-  return <SearchOptionsComponent />;
-}
